test(orchestration): split phase output coverage by owner
